+++ contracts/sdk/AQueryProofExecutor.sol
@@ -83,7 +83,7 @@ abstract contract AQueryProofExecutor is Initializable {
         bytes32 registrationRoot_,
         uint256 currentDate_,
         bytes memory userPayload_
-    ) public virtual returns (uint256 builder_);
+    ) public view virtual returns (uint256 builder_);
 
     /**
      * @notice Executes the full ZK proof verification workflow for a Circom (Groth16) proof.
@@ -145,7 +145,7 @@ abstract contract AQueryProofExecutor is Initializable {
         bytes32 registrationRoot_,
         uint256 currentDate_,
         bytes memory userPayload_
-    ) public virtual returns (bytes32[] memory publicSignals) {
+    ) public view virtual returns (bytes32[] memory publicSignals) {
         uint256 builder_ = _buildPublicSignals(registrationRoot_, currentDate_, userPayload_);
         builder_.withIdStateRoot(registrationRoot_);